scheduler/requirements.txt
apscheduler~=3.10.4
redis~=5.0.3
fastapi~=0.110.0
pandas~=2.2.1
requests~=2.31.0
bs4~=0.0.2
beautifulsoup4~=4.12.3
python-dotenv~=1.0.1
uvicorn~=0.28.0
aiohttp~=3.9.3
colorlog~=6.8.2
confluent-kafka~=2.3.0
pydantic~=2.6.4
joblib~=1.3.2
scikit-learn